name changed for search result
